+++ src/Domain/Entities/CountryData.cs
@@ -4,5 +4,5 @@ public class CountryData
   public int CountryId { get; set; }
   public string CountryName { get; set; } = string.Empty;
 
-  public Dictionary<string, int> YearData { get; set; } = new();
+  public Dictionary<string, decimal> YearData { get; set; } = new();
 }
